ENH: implement image building using packer
packer is a tool by hashicorp (who also make vagrant) to automate the
creation of images for virtual machines.
This commit introduces a packer configuration file for a Debian 10 based
virtual machine. Together with this configuration file, a Debian preseed
configuration is located in the http folder, which will be served to the
Debian installer by packer. After the installation process packer logs
into the running machine via ssh and runs the vagrant recipes.
The commit also introduced a .gitignore file to ignore packer files.
This file also lists the GeoLite2-City.mmdb file, so that it is not
committed by accident.